requirements.txt
streamlit==1.32.0
pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
tensorflow-cpu==2.15.0
tensorflow-hub==0.15.0
numpy==1.24.3
protobuf==4.25.1 